requirements.txt
Flask==3.1.1
requests==2.32.3
psutil==7.0.0
PyNaCl==1.5.0 